feat: update i3.sh

+++ bash-scripts/i3.sh
@@ -12,15 +12,30 @@ do
             ;;
         "logout")
             i3-msg exit
+            read -p "Are you sure you want to logout? (y/n) " -n 1 -r
+            if [[ $REPLY =~ ^[Yy]$ ]]
+              i3-msg exit
+            fi
             ;;
         "suspend")
             lock && systemctl suspend
+            read -p "Are you sure you want to suspend? (y/n) " -n 1 -r
+            if [[ $REPLY =~ ^[Yy]$ ]]
+              lock && systemctl suspend
+            fi
             ;;
         "reboot")
             systemctl reboot
+            read -p "Are you sure you want to reboot? (y/n) " -n 1 -r
+            if [[ $REPLY =~ ^[Yy]$ ]]
+              systemctl reboot
+            fi
             ;;
         "shutdown")
+          read -p "Are you sure you want to shutdown? (y/n) " -n 1 -r
+          if [[ $REPLY =~ ^[Yy]$ ]]
             systemctl poweroff
+          fi
             ;;
         "cancel")
           exit 0
